feat:implement the design for desktop
Download all the images, create the basic structure in JSX. Then, start
to add styles with tailwind classes. Some styles needed custom styles, i
create classes in the index.css for them. Create a component for the
task components and map an array of values to display similar to the
design for desktop