feat:implement delete TODO feature
Use React Hooks to manage the state and modify the data added by the
user. Change the value list format to use id for seach and filter TODOs.
Add a fade-out animation to smoothly show every delete to the user and
wait until the animation is finished to delete the TODO