feat:implement edit TODO feature
Add new icons for the user to select his options. Make an Input
components to abstract a bit more the responsibilities. Show an input
with the new icons to the user to get the new input. Manage the events
and state to replace the card title if the user selects "confirm" or the
user selects cancel. Make some functions in app component to abstract
the event functionalities and improve readability